pyproject.toml
[project]
name = "opendb"
version = "1.3.0"
description = "The AI-native file database and memory store. Built for LLM agents to read, search, and remember."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "wuwangzhang1216", email = "wuwangzhang1216@users.noreply.github.com"}
]
maintainers = [
    {name = "wuwangzhang1216", email = "wuwangzhang1216@users.noreply.github.com"}
]
keywords = ["file", "parser", "pdf", "docx", "pptx", "xlsx", "ocr", "ai", "agents", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.29.0",
    "python-multipart>=0.0.9",
    "PyMuPDF>=1.24.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "python-magic-bin>=0.4.14; sys_platform == 'win32'",
    "python-magic>=0.4.27; sys_platform != 'win32'",
    "pydantic-settings>=2.0.0",
    "watchdog>=4.0.0",
    "jieba>=0.42.1",
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.0.0",
    "httpx>=0.28.0",
]
integration = [
    "httpx>=0.28.0",
]
# Embedded mode: SQLite backend + CLI + vision
embedded = [
    "aiosqlite>=0.19.0",
    "httpx>=0.27.0",
]
cli = [
    "aiosqlite>=0.19.0",
    "typer>=0.12.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
    "aiosqlite>=0.19.0",
]

[project.urls]
Homepage = "https://github.com/wuwangzhang1216/openDB"
Repository = "https://github.com/wuwangzhang1216/openDB"
Issues = "https://github.com/wuwangzhang1216/openDB/issues"
Documentation = "https://github.com/wuwangzhang1216/openDB#readme"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
opendb = "opendb.cli:main"

[tool.setuptools.packages.find]
include = ["opendb_core*", "mcp_server*", "opendb_integration*", "opendb*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
